+++ etc/mepo-completion.bash
@@ -1,13 +1,8 @@
 #!/usr/bin/env bash
 
-# complete -W "init clone status checkout branch diff where whereis history" mepo
-
-# SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
-SCRIPT_DIR=$(dirname $(realpath $0))
-
 _get_mepo_commands() {
     local mepo_cmd_list=""
-    local mepo_dir=$(python3 $SCRIPT_DIR/mepo-path.py)
+    local mepo_dir=$(mepo --location)
     for pyfile in $(ls ${mepo_dir}/command/*.py*); do
 	command=${pyfile##*/} # remove path
 	command=${command%.*} # remove extension