+++ src/mepo/cmdline/parser.py
@@ -1,4 +1,5 @@
 import argparse
+import warnings
 
 from .branch_parser import MepoBranchArgParser
 from .stash_parser import MepoStashArgParser
@@ -14,6 +15,19 @@ def get_version():
     return metadata.version("mepo")
 
 
+class LocationAction(argparse._StoreTrueAction):
+
+    def __init__(self, option_strings, dest, const=True, help=None):
+        super().__init__(option_strings, dest, const, help=help)
+
+    def __call__(self, parser, namespace, values, option_string=None):
+        import os, sys
+        import mepo
+
+        print(os.path.dirname(mepo.__file__))
+        sys.exit(0)
+
+
 class MepoArgParser:
 
     __slots__ = ["parser", "subparsers"]
@@ -23,6 +37,9 @@ class MepoArgParser:
             description="Tool to manage (m)ultiple r(epo)s"
         )
         self.parser.add_argument("--version", action="version", version=get_version())
+        self.parser.add_argument(
+            "--location", action=LocationAction, help=argparse.SUPPRESS
+        )
         self.subparsers = self.parser.add_subparsers()
         self.subparsers.title = "mepo commands"
         self.subparsers.required = True
@@ -58,6 +75,9 @@ class MepoArgParser:
         return self.parser.parse_args()
 
     def __init(self):
+        warnings.warn(
+            "init will be removed in version 3, use clone instead", DeprecationWarning
+        )
         init = self.subparsers.add_parser(
             "init",
             description="Initialize mepo based on `config-file`",
@@ -85,7 +105,7 @@ class MepoArgParser:
             aliases=mepoconfig.get_command_alias("clone"),
         )
         clone.add_argument(
-            "repo_url", metavar="URL", nargs="?", default=None, help="URL to clone"
+            "url", metavar="URL", nargs="?", default=None, help="URL to clone"
         )
         clone.add_argument(
             "directory",
@@ -106,7 +126,7 @@ class MepoArgParser:
             metavar="registry",
             nargs="?",
             default=None,
-            help="Registry (default: components.yaml)",
+            help="Registry (default: %(default)s)",
         )
         clone.add_argument(
             "--style",
@@ -126,8 +146,18 @@ class MepoArgParser:
             metavar="partial-type",
             nargs="?",
             default=None,
-            choices=["off", "blobless", "treeless"],
-            help='Style of partial clone, default: None, allowed options: %(choices)s. Off means a "normal" full git clone, blobless means cloning with "--filter=blob:none" and treeless means cloning with "--filter=tree:0". NOTE: We do *not* recommend using "treeless" as it is very aggressive and will cause problems with many git commands.',
+            choices=[None, "blobless", "treeless"],
+            help=(
+                """
+                Style of partial clone, default: %(default)s.
+                Allowed options: %(choices)s.
+                None: normal full git clone,
+                blobless: cloning with "--filter=blob:none",
+                treeless: cloning with "--filter=tree:0".
+                NOTE: We do *not* recommend using "treeless" as it is very
+                aggressive and will cause problems with many git commands.
+                """
+            ),
         )
 
     def __list(self):